Create Censor
This script will take a phrase as input and then take a second string as input to be censored out of the original phrase. So if the first string were "Hello World" and you gave "l" as the string to be censored out. The script would return "He**o Wor*d"