Create Hearts.py
This is a program that will work as a score keeper for you when you are playing hearts with some friends. it has error checking build in for when people "accidentally" miscount how many hearts they might have taken on any given round. it also gives you instructions on which  direction you need to pass your cards at the beginning of each round so you dont forget. This program will also automatically calculate if someone shot the moon. At the end of the game it will print out the results of scores and let everyone know what place people got.